# back-furniture/src/api/swagger.yaml
openapi: 3.0.0
info:
  title: Furniture Store
  description: This is Furniture Shop Backend 
  version: 0.1.9
servers:
  - url: /api/v1
security:
  - BearerAuth: []
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: Enter your Bearer token here
consumes:
  - application/json 
paths:
  /users:
    get:
      tags:
        - "User"
      x-swagger-router-controller: user
      operationId: getUser
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML.
      parameters:
        - name: currentPage
          in: query
          description: currentPage
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          description: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          description: "sort column"
          required: false
          schema:
            type: integer
        - name: sortDirection
          in: query
          description: "sort i.e asc desc"
          required: false
          schema:
            type: integer
            enum: 
              - asc
              - desc                                   
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Users"
    post: 
      x-swagger-router-controller: user
      operationId: addUser
      summary: Register a new user.
      tags:
        - "User"      
      description: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: sumit
                lastName:
                  type: string
                  example: kumar
                username:
                  type: string
                  example: sumit1234
                email:
                  type: string
                  example: sumitbhatia0796@gmail.com
                password:
                  type: string
                  example: sumit9876
                mobile:
                  type: string
                  example: "7807465496"
                  pattern: "^[0-9()+][0-9-() ]*[0-9]$"
                  maxLength: 30
                dataOfBirth:
                  type: string
                  example: 11-12-2001
                gender:
                  type: string
                  example: male
                address:
                  type: string
                  example: vill khaterwar teh bhoranj
                country:
                  type: string
                  example: india
                state:
                  type: string
                  example: Himachal Pardesh
                city:
                  type: string
                  example: Hamirpur
                pin:
                  type: string 
                  example: 176045 
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"      
    delete:
      tags:
        - "User"
      x-swagger-router-controller: user
      operationId: deleteAllUser
      summary: Delete All Available User.
      responses:
        '204':    # status code
          description: Default Response
  /users/{userId}:
    get:
      tags:
        - "User"
      x-swagger-router-controller: user
      operationId: getUserById
      summary: Returns  user.
      description: Optional extended description in CommonMark or HTML.
      parameters:
        - name: userId
          in: path
          description: userId
          required: false
          schema:
            type: string
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/User"
      security:
        - authentication: []             
    patch: 
      x-swagger-router-controller: user
      operationId: updateUserById
      summary: Register a new user.
      tags:
        - "User"      
      description: Register a new user
      parameters:
        - name: userId
          in: path
          description: userId
          required: false
          schema:
            type: string      
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: sumit9876
                mobile:
                  type: string
                  example: "7807465496"
                  pattern: "^[0-9()+][0-9-() ]*[0-9]$"
                  maxLength: 30
                address:
                  type: string
                  example: vill khaterwar teh bhoranj
                country:
                  type: string
                  example: india
                state:
                  type: string
                  example: Himachal Pardesh
                city:
                  type: string
                  example: Hamirpur
                pin:
                  type: string 
                  example: 176045 
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"      
    delete:
      tags:
        - "User"
      x-swagger-router-controller: user
      operationId: deleteUserById
      summary: Delete All Available User.
      parameters:
        - name: userId
          in: path
          description: userId
          required: false
          schema:
            type: string        
      responses:
        '204':    # status code
          description: Default Response
  /users/login:          
    post: 
      x-swagger-router-controller: user
      operationId: loginUser
      summary: Register a new user.
      tags:
        - "User"      
      description: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: sumit1234
                password:
                  type: string
                  example: sumit9876
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"          
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: Enter your Bearer token here
  schemas:
    User:
      type: object
      properties:
        firstName:
          type: string
          example: Anil      
        lastName:
          type: string
          example: kumar 
        username:
          type: string
          example: Anil34 
        email:
          type: string
          example: anilbhatia34@gmail.com 
        mobile:
          type: string
          example: Anil 
        gender:
          type: string
          example: Anil 
        country:
          type: string
          example: Anil 
        city:
          type: string
          example: Anil 
        address:
          type: string
          example: Anil    
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User" 
...
    Token:
      type: object
      properties:
        name:
          type: string
          example: Anil 
        token:
          type: string
          example: wwqewqereffssadweweewrasdsadasd    
...